optimized install.sh to be more idempotent

+++ setup/install.sh
@@ -11,22 +11,31 @@ set -e
 apt-get install -y python-software-properties
 
 # ElasticSearch
-wget -nv -O - http://packages.elasticsearch.org/GPG-KEY-elasticsearch | apt-key add -
-echo "deb http://packages.elasticsearch.org/elasticsearch/1.0/debian stable main" > /etc/apt/sources.list.d/elasticsearch.list
+if [ ! -e /etc/apt/sources.list.d/elasticsearch.list ]; then
+    wget -nv -O - http://packages.elasticsearch.org/GPG-KEY-elasticsearch | apt-key add -
+    echo "deb http://packages.elasticsearch.org/elasticsearch/1.0/debian stable main" > /etc/apt/sources.list.d/elasticsearch.list
+fi
 
 # PHP Stable
-add-apt-repository -y ppa:ondrej/php5
+[ -e "/etc/apt/sources.list.d/ondrej-php5-$(lsb_release -cs).list" ] || add-apt-repository -y ppa:ondrej/php5
 
 # PHP Composer
-add-apt-repository -y ppa:duggan/composer
+[ -e "/etc/apt/sources.list.d/duggan-composer-$(lsb_release -cs).list" ] || add-apt-repository -y ppa:duggan/composer
 
 # NodeJs
-add-apt-repository -y ppa:chris-lea/node.js
+[ -e "/etc/apt/sources.list.d/chris-lea-node_js-$(lsb_release -cs).list" ] || add-apt-repository -y ppa:chris-lea/node.js
 
 # Percona
-apt-key adv --keyserver keys.gnupg.net --recv-keys 1C4CBDCDCD2EFD2A
-echo "deb http://repo.percona.com/apt $(lsb_release -cs) main" > /etc/apt/sources.list.d/percona.list
-echo "deb-src http://repo.percona.com/apt $(lsb_release -cs) main" >> /etc/apt/sources.list.d/percona.list
+if [ ! -e /etc/apt/sources.list.d/percona.list ]; then
+    apt-key adv --keyserver keys.gnupg.net --recv-keys 1C4CBDCDCD2EFD2A
+    echo "deb http://repo.percona.com/apt $(lsb_release -cs) main" > /etc/apt/sources.list.d/percona.list
+    echo "deb-src http://repo.percona.com/apt $(lsb_release -cs) main" >> /etc/apt/sources.list.d/percona.list
+fi
+
+echo "deb mirror://mirrors.ubuntu.com/mirrors.txt $(lsb_release -cs)           main restricted universe multiverse" >  /etc/apt/sources.list
+echo "deb mirror://mirrors.ubuntu.com/mirrors.txt $(lsb_release -cs)-updates   main restricted universe multiverse" >> /etc/apt/sources.list
+echo "deb mirror://mirrors.ubuntu.com/mirrors.txt $(lsb_release -cs)-backports main restricted universe multiverse" >> /etc/apt/sources.list
+echo "deb mirror://mirrors.ubuntu.com/mirrors.txt $(lsb_release -cs)-security  main restricted universe multiverse" >> /etc/apt/sources.list
 
 # http://serverfault.com/questions/322400/install-mysql-server-and-set-password-from-the-command-line
 # /var/lib/dpkg/info/percona-server-server-5.6.config
@@ -53,17 +62,16 @@ echo 'dbconfig-common dbconfig-common/app-password-confirm password root' | debc
 echo 'dbconfig-common dbconfig-common/password-confirm password root' | debconf-set-selections
 
 
+apt-get update
+aptitude dist-upgrade -y
 
 ######
 # Install
 ######
-apt-get update
-
-aptitude dist-upgrade -y
 
 cat "/media/vagrant/setup/packages.list" | xargs aptitude install -y linux-headers-`uname -r`
 
-npm install -g bower
+npm install -g bower less
 
 rsync -av "/media/vagrant/setup/files/" /
 
@@ -81,7 +89,7 @@ sed -i "s/www-data/vagrant/" /etc/php5/fpm/pool.d/www.conf
 for module in rewrite alias actions vhost_alias setenvif proxy proxy_http; do
   a2enmod $module
 done
-a2ensite wmc.conf
+a2ensite wmc
 sed -i "s/www-data/vagrant/" /etc/apache2/envvars
 chown -R vagrant:vagrant /var/lib/apache2/
 
@@ -96,9 +104,9 @@ ln -sfv /media/data/projects /home/vagrant/projects
 
 # PEAR
 pear -q upgrade pear
-pear -q channel-discover pear.phpunit.de
-pear -q channel-discover components.ez.no
-pear -q channel-discover pear.symfony.com
+pear list-channels | grep -qF pear.phpunit.de || pear -q channel-discover pear.phpunit.de
+pear list-channels | grep -qF components.ez.no || pear -q channel-discover components.ez.no
+pear list-channels | grep -qF pear.symfony.com || pear -q channel-discover pear.symfony.com
 
 service php5-fpm restart
 service apache2 restart
